+++ razorboard/src/app/http-events.service.ts
@@ -4,14 +4,19 @@ import { Subject } from 'rxjs';
 @Injectable({
   providedIn: 'root'
 })
-export class HttpLoadingService {
+export class HttpEventsService {
 
   constructor() { }
   loading = new Subject<boolean>();
+  status$ = new Subject<number>();
+
   show() {
     this.loading.next(true);
   }
   hide() {
     this.loading.next(false);
+  }
+  statusNotify(value: number) {
+    this.status$.next(value);
   }
 }