+++ razorboard/src/app/http-loading.interceptor.spec.ts
@@ -4,23 +4,23 @@ import {
   HttpTestingController,
 } from '@angular/common/http/testing';
 import { RazorapiService } from './razorapi.service';
-import { HttpLoadingService } from './http-loading.service';
+import { HttpEventsService } from './http-events.service';
 import { HttpLoadingInterceptor } from './http-loading.interceptor';
 import { HTTP_INTERCEPTORS } from '@angular/common/http';
 
-import { HttpLoadingServiceStub } from '../testing/http-loading.service.stub';
+import { HttpEventsServiceStub } from '../testing/http-events.service.stub';
 
 describe(`HttpLoadingInterceptor`, () => {
   let service: RazorapiService;
   let httpMock: HttpTestingController;
-  const httpLoadingStub = new HttpLoadingServiceStub();
+  const httpEventsStub = new HttpEventsServiceStub();
 
   beforeEach(() => {
     TestBed.configureTestingModule({
       imports: [HttpClientTestingModule],
       providers: [
         RazorapiService,
-        { provide: HttpLoadingService, useValue: httpLoadingStub },
+        { provide: HttpEventsService, useValue: httpEventsStub },
         {
           provide: HTTP_INTERCEPTORS,
           useClass: HttpLoadingInterceptor,
@@ -37,6 +37,6 @@ describe(`HttpLoadingInterceptor`, () => {
   it('should manage loading service', fakeAsync(() => {
     service.getNodes().subscribe(response => { }).unsubscribe();
     tick();
-    expect(httpLoadingStub.calls).toEqual(['show', 'hide']);
+    expect(httpEventsStub.calls).toEqual(['show', 'hide']);
   }));
 });